feat: Add maintainerr
This image is frustratingly coupled to the UID 1000. NixOS docker
options have no way to set `--user=${user.uid}` because the `uid` is not
known unless explicitly set (which I would prefer to avoid). I've tried
all sorts of workarounds by modifying the `systemd.services.*` entry to
calculate the UID and set it on `docker run` but because of shell
escaping on `.extraOptions` I can't get it to work.

For now, I'm just overriding this and forcing it to be 1000 but it's
a pretty inelegant solution, especially since `mkArr` exposes
`user`/`group` options that will be inert.

I may need to break this out from the general `mkArr` structure _or_
look into building my own image which is _derived_ from upstream, which
adds Docker `ARG`s to take (and use in `USER` directive) `PUID`/`PGID`
ENV vars like LinuxServer.io images.