fix: switch back to depending on automount
I forgot that I configured the NFS mount to unmount after inactivity,
which brings the services down. Depending on the automount is sufficient
to re-activate the mount on-demand if the service tries to write.

Ideally, having the mount itself crash/fail would bring down the
services, but not having it be idle/unmounted