refactor: change NFS mount point for NAS + mount temporary NAS
This also creates a MergerFS mountpoint to combine both, but this is
still WIP. Trying to use it as a path in Docker images creates
permissions errors. I'm not sure why since from within the docker image,
the UID=1024/GID=100 is displayed when using NFS directly and this
appears to be the case under MergerFS, but at least the Plex container
complains about not being able to write to the latter (even though
I objectively can `touch` a file while running in the container). It
seems it is examining the owner of directory, but it is shows as 1024
regardless.

I'm guessing there is some NFS-aware handling for permission squashing
which isn't being used in context of the MergerFS mount. More spelunking
necessary...